requirements.txt
aiogram~=2.22.1
requests~=2.28.1
python-dotenv~=0.21.0
beautifulsoup4~=4.11.1
pdf2image~=1.16.0
pandas~=1.5.1
aiohttp~=3.8.3
openpyxl~=3.0.10